修复 get_maya_main_window 方法中的类型转换错误，将 long 类型改为 int 类型

+++ mayaPlugins/scripts/doodle/main.py
@@ -17,7 +17,7 @@ class main(object):
   @staticmethod
   def get_maya_main_window():
     main_window_ptr = omui.MQtUtil.mainWindow()
-    return wrapInstance(long(main_window_ptr), QWidget)
+    return wrapInstance(int(main_window_ptr), QWidget)
 
   @staticmethod
   def add_doodle_tool_widget(*args):
